+++ src/com/intellij/idea/plugin/hybris/acl/editor/AclInEditorResultsView.kt
@@ -20,37 +20,37 @@ package com.intellij.idea.plugin.hybris.acl.editor
 
 import com.intellij.idea.plugin.hybris.editor.InEditorResultsView
 import com.intellij.idea.plugin.hybris.tools.remote.execution.DefaultExecutionResult
-import com.intellij.idea.plugin.hybris.ui.Dsl
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.ui.DialogPanel
 import com.intellij.ui.EditorNotificationPanel
 import com.intellij.ui.InlineBanner
 import com.intellij.ui.dsl.builder.Align
 import com.intellij.ui.dsl.builder.Panel
 import com.intellij.ui.dsl.builder.TopGap
-import com.intellij.ui.dsl.builder.panel
 import com.intellij.ui.dsl.gridLayout.UnscaledGaps
-import com.intellij.util.ui.JBUI
 import kotlinx.coroutines.CoroutineScope
-import java.awt.Dimension
-import javax.swing.ScrollPaneConstants
 
 @Service(Service.Level.PROJECT)
 class AclInEditorResultsView(project: Project, coroutineScope: CoroutineScope) : InEditorResultsView<AclSplitEditor, DefaultExecutionResult>(project, coroutineScope) {
 
-    override suspend fun render(fileEditor: AclSplitEditor, result: DefaultExecutionResult) = panel {
-        when {
-            result.hasError -> errorView(result.errorMessage ?: "An error was encountered while processing the request.", result.errorDetailMessage)
-            result.output != null -> resultsView(result.output)
-            else -> noResultsView()
+    override suspend fun render(fileEditor: AclSplitEditor, results: Collection<DefaultExecutionResult>): DialogPanel = results.firstOrNull()
+        .takeIf { results.size == 1 }
+        ?.let { result ->
+            panelView {
+                when {
+                    result.hasError -> it.errorView(
+                        result.errorMessage ?: "An error was encountered while processing the request.",
+                        result.errorDetailMessage
+                    )
+
+                    result.output != null -> it.resultsView(result.output)
+                    else -> it.noResultsView()
+                }
+            }
         }
-    }
-        .apply { border = JBUI.Borders.empty(5, 16, 10, 16) }
-        .let { Dsl.scrollPanel(it, ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER) }
-        .apply {
-            minimumSize = Dimension(minimumSize.width, 150)
-        }
+        ?: multiResultsNotSupportedView()
 
     private fun Panel.resultsView(output: String) {
         panel {