+++ src/com/intellij/idea/plugin/hybris/groovy/editor/GroovyInEditorResultsView.kt
@@ -20,44 +20,94 @@ package com.intellij.idea.plugin.hybris.groovy.editor
 
 import com.intellij.idea.plugin.hybris.editor.InEditorResultsView
 import com.intellij.idea.plugin.hybris.tools.remote.execution.DefaultExecutionResult
-import com.intellij.idea.plugin.hybris.ui.Dsl
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
+import com.intellij.ui.EditorNotificationPanel
+import com.intellij.ui.InlineBanner
 import com.intellij.ui.dsl.builder.Align
 import com.intellij.ui.dsl.builder.Panel
-import com.intellij.ui.dsl.builder.panel
-import com.intellij.util.ui.JBUI
+import com.intellij.ui.dsl.builder.TopGap
+import com.intellij.ui.dsl.gridLayout.UnscaledGaps
 import kotlinx.coroutines.CoroutineScope
-import java.awt.Dimension
 import java.lang.Boolean
+import javax.swing.JComponent
 import javax.swing.JEditorPane
-import javax.swing.ScrollPaneConstants
 import kotlin.String
 import kotlin.apply
 import kotlin.let
+import kotlin.takeIf
 
 @Service(Service.Level.PROJECT)
 class GroovyInEditorResultsView(project: Project, coroutineScope: CoroutineScope) : InEditorResultsView<GroovySplitEditor, DefaultExecutionResult>(project, coroutineScope) {
 
-    override suspend fun render(fileEditor: GroovySplitEditor, result: DefaultExecutionResult) = panel {
-        when {
-            result.hasError -> errorView(result.errorMessage ?: "An error was encountered while processing the request.", result.errorDetailMessage)
-            result.result != null || result.output != null -> {
-                if (result.result != null) group("Result", result.result)
-                if (result.output != null) group("Output", result.output)
+    override suspend fun render(fileEditor: GroovySplitEditor, results: Collection<DefaultExecutionResult>): JComponent {
+        return results.firstOrNull()
+            .takeIf { results.size == 1 }
+            ?.let { result ->
+                panelView {
+                    when {
+                        result.hasError -> it.errorView(
+                            result.errorMessage ?: "An error was encountered while processing the request.",
+                            result.errorDetailMessage
+                        )
+
+                        result.result != null || result.output != null -> {
+                            if (result.result != null) it.group("Result", result.result)
+                            if (result.output != null) it.group("Output", result.output)
+                        }
+
+                        else -> it.noResultsView()
+                    }
+                }
             }
+            ?: panelView {
+                val resultsWithErrors = results.count { result -> result.hasError }
 
-            else -> noResultsView()
-        }
+                if (resultsWithErrors > 0) {
+                    it.panel {
+                        row {
+                            cell(
+                                InlineBanner(
+                                    """
+                                        Groovy script execution resulted to an error on $resultsWithErrors of ${results.size} replicas.<br>
+                                        Details of each individual execution result can be found below.
+                                        """.trimIndent(),
+                                    EditorNotificationPanel.Status.Warning,
+                                ).showCloseButton(false)
+                            )
+                                .align(Align.FILL)
+                                .resizableColumn()
+                        }.topGap(TopGap.SMALL)
+                    }
+                        .customize(UnscaledGaps(16, 16, 16, 16))
+                }
+
+                results
+                    .sortedBy { result -> result.replicaContext?.replicaId }
+                    .forEach { result ->
+                        it.collapsibleGroup("Replica: ${result.replicaContext?.replicaId ?: ""}") {
+                            when {
+                                result.hasError -> errorView(
+                                    result.errorMessage ?: "An error was encountered while processing the request.",
+                                    result.errorDetailMessage
+                                )
+
+                                result.result != null || result.output != null -> {
+                                    group("Result", result.result)
+                                    group("Output", result.output)
+                                }
+
+                                else -> noResultsView()
+                            }
+                        }.expanded = true
+                    }
+            }
     }
-        .apply { border = JBUI.Borders.empty(5, 16, 10, 16) }
-        .let { Dsl.scrollPanel(it, ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER) }
-        .apply {
-            minimumSize = Dimension(minimumSize.width, 150)
-        }
 
-    private fun Panel.group(title: String, text: String) {
+    private fun Panel.group(title: String, text: String?) {
+        if (text == null) return
+
         collapsibleGroup(title) {
             row {
                 cell(