+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExValidateAction.kt
@@ -50,7 +50,7 @@ class ImpExValidateAction : ExecuteStatementAction<HybrisImpexConsole, ImpExSpli
 
         if (fileEditor?.inEditorResults ?: false) {
             fileEditor.putUserData(KEY_QUERY_EXECUTING, true)
-            fileEditor.showLoader()
+            fileEditor.showLoader(context)
 
             ImpExExecutionClient.getInstance(project).execute(context) { coroutineScope, result ->
                 fileEditor.renderExecutionResult(result)