+++ src/com/intellij/idea/plugin/hybris/tools/remote/execution/logging/LoggingExecutionContext.kt
@@ -23,7 +23,7 @@ import com.intellij.idea.plugin.hybris.tools.remote.execution.ExecutionContext
 import com.intellij.idea.plugin.hybris.tools.remote.http.HybrisHacHttpClient
 
 data class LoggingExecutionContext(
-    override val title: String,
+    override val executionTitle: String,
     private val loggerName: String,
     private val logLevel: LogLevel,
     val timeout: Int = HybrisHacHttpClient.DEFAULT_HAC_TIMEOUT,