+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 ### `Groovy` enhancements
 - Introduced `In-Editor Results` view [#1477](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1477)
 - Introduced `batch` execution mode for Groovy files [#1450](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1450)
+- Show replica-specific results within `In-Editor Results` view [#1478](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1478)
 
 ### `Remote Connection` enhancements
 - Shorten CCv2 url without custom name [#1449](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1449)